fix: type safety and error handling improvements
## ai-complete-handler.ts
- Replace  type with proper typed interface for aiSettings
- Ensures compiler enforces correct types and prevents accidental properties

## ai-stream-handler.ts
- Replace  type with proper typed interface for aiSettings
- Improve partial content error handling: persist partial content when streaming fails
- Placeholder is now updated with partial content before yielding error to client

## ai-database-helpers.test.ts
- Fix TDZ errors by using vi.hoisted() before vi.mock() calls
- Remove unused imports (eq, and from drizzle-orm)

## ai-database-helpers.ts
- Add isValidProvider type guard function to validate AIProvider strings
- Remove `as any` cast - use runtime validation with type guard instead
- Fix duplicate detection query to use desc(createdAt) for most recent message check

## ai-schemas.ts & ai-schemas.test.ts
- Change maxTokens and topK to use .int() - only accept integers
- Update tests to assert decimal inputs are rejected

+++ packages/api/src/lib/ai-schemas.ts
@@ -85,7 +85,7 @@ export const completionSettingsSchema = z.object({
    * Maximum number of tokens to generate (1-1000000)
    * Limits the length of the model's response
    */
-  maxTokens: z.number().min(1).max(1000000).optional(),
+  maxTokens: z.number().int().min(1).max(1000000).optional(),
 
   /**
    * Nucleus sampling parameter (0-1)
@@ -99,7 +99,7 @@ export const completionSettingsSchema = z.object({
    * Limits token selection to the k most probable tokens
    * 0 = disabled, higher values = more diverse options
    */
-  topK: z.number().min(0).max(100).optional(),
+  topK: z.number().int().min(0).max(100).optional(),
 
   /**
    * Frequency penalty (-2 to 2)
